requirements.txt
Django==6.0.4
psycopg2-binary==2.9.9
whitenoise==6.7.0
dj-database-url==2.2.0
python-dotenv==1.0.1
gunicorn==22.0.0